requirements.txt
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
configobj==5.0.6
cryptography==36.0.1
# dbus-python==1.2.4
decorator==4.2.1
futures==2.2.0
# google-auth==2.3.3
# gpg==1.13.1
grpcio==1.43.0
grpcio-tools==1.43.0
html5lib==0.999999999
idna==3.3
# isc==2.0
jwt==1.3.1
kubernetes==21.7.0
lxml==4.2.3
# nftables==0.1
oauthlib==3.1.1
# perf==0.1
pexpect==4.3.1
ply==3.9
protobuf==3.19.3
psutil==5.4.3
ptyprocess==0.5.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.3
pycparser==2.21
pydbus==0.6.0
# PyGObject==3.28.3
pyparsing==2.1.10
python-dateutil==2.6.1
# python-linux-procfs==0.6.3
pyudev==0.21.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
# rpm==4.14.3
rsa==4.8
# schedutils==0.6
# selinux==2.9
semver==2.13.0
# sepolicy==1.1
# setools==4.3.0
# setroubleshoot==1.1
# six==1.11.0
# slip==0.6.4
# slip.dbus==0.6.4
# sos==4.0
spython==0.1.17
# SSSDConfig==2.4.0
# syspurpose==1.28.13
# systemd-python==234
urllib3==1.26.8
webencodings==0.5.1
# websocket-client==1.2.3
